+++ plugins/api-docs/src/components/ApiExplorerPage/DefaultApiExplorerPage.tsx
@@ -40,6 +40,7 @@ import React from 'react';
 import { registerComponentRouteRef } from '../../routes';
 
 const defaultColumns: TableColumn<CatalogTableRow>[] = [
+  CatalogTable.columns.createTitleColumn({ hidden: true }),
   CatalogTable.columns.createNameColumn({ defaultKind: 'API' }),
   CatalogTable.columns.createSystemColumn(),
   CatalogTable.columns.createOwnerColumn(),