+++ plugins/home/src/index.ts
@@ -31,9 +31,10 @@ export {
   ComponentTabs,
   ComponentTab,
   WelcomeTitle,
+  HeaderWorldClock,
 } from './plugin';
-export { SettingsModal, HeaderWorldClock } from './components';
+export { SettingsModal } from './components';
 export * from './assets';
-export type { ClockConfig } from './components';
+export type { ClockConfig } from './homePageComponents';
 export { createCardExtension } from './extensions';
 export type { ComponentRenderer } from './extensions';