+++ plugins/search/src/components/SearchModal/SearchModal.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import {
   Dialog,
   DialogActions,
@@ -35,6 +35,7 @@ import {
   SearchResult,
   SearchResultPager,
   useSearch,
+  useSearchContextCheck,
 } from '@backstage/plugin-search-react';
 import { useRouteRef } from '@backstage/core-plugin-api';
 import { Link, useContent } from '@backstage/core-components';
@@ -170,6 +171,15 @@ export const Modal = ({ toggleModal }: SearchModalProps) => {
   );
 };
 
+const Context = ({ children }: PropsWithChildren<{}>) => {
+  // Checks if there is a parent context already defined and, if not, creates a new local context.
+  const hasParentContext = useSearchContextCheck();
+  if (hasParentContext) {
+    return <>{children}</>;
+  }
+  return <SearchContextProvider>{children}</SearchContextProvider>;
+};
+
 /**
  * @public
  */
@@ -194,11 +204,11 @@ export const SearchModal = ({
       hidden={hidden}
     >
       {open && (
-        <SearchContextProvider>
+        <Context>
           {(children && children({ toggleModal })) ?? (
             <Modal toggleModal={toggleModal} />
           )}
-        </SearchContextProvider>
+        </Context>
       )}
     </Dialog>
   );